refact: DIP, ISP에 따라 리팩토링
홈과 디테일 레포와 소스를 분리한 후 각각의 인터페이스를 만들어 참조하게 함
