+++ app/src/main/java/com/example/potatoservice/model/remote/ActivityDataClass.kt
@@ -62,7 +62,39 @@ data class ActivityDetail(
     val url: String,
     val category: String,
     val institute: Institute
-)
+){
+    companion object{
+        val nullActivityDetail = ActivityDetail(
+            0,
+            "봉사 활동 제목 정보 없음",
+            "봉사 활동 장소 정보 없음",
+            "봉사 활동 설명 정보 없음",
+            "공지 시작 날짜 정보 없음",
+            "공지 종료 날짜 정보 없음",
+            "봉사 활동 시작 날짜 정보 없음",
+            "봉사 활동 종료 날짜 정보 없음",
+            0,
+            0,
+            0,
+            false,
+            false,
+            false,
+            0,
+            "봉사 활동 담당자 정보 없음",
+            "봉사 활동 전화 번호 정보 없음",
+            "봉사 활동 url 정보 없음",
+            "카테고리 정보 없음",
+            Institute(
+                0,
+                "기관 정보 없음",
+                "기관 주소 정보 없음",
+                0.0,
+                0.0,
+                "기관 전화 번호 정보 없음"
+            )
+        )
+    }
+}
 
 data class Institute(
     val instituteId:Int,
