+++ app/src/main/java/com/example/potatoservice/module/RepositoryModule.kt
@@ -1,8 +1,13 @@
 package com.example.potatoservice.module
 
 import com.example.potatoservice.model.APIService
+import com.example.potatoservice.ui.detail.DetailRepository
+import com.example.potatoservice.ui.detail.DetailSearchData
+import com.example.potatoservice.ui.detail.DetailSearchDataSource
 import com.example.potatoservice.ui.home.HomeRepository
-import com.example.potatoservice.ui.home.RemoteDataSource
+import com.example.potatoservice.ui.home.HomeSearchData
+import com.example.potatoservice.ui.home.HomeSearchDataSource
+import dagger.Binds
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
@@ -14,13 +19,14 @@ import javax.inject.Singleton
 object RepositoryModule {
 	@Provides
 	@Singleton
-	fun provideRemoteDataSource(apiService: APIService): RemoteDataSource {
-		return RemoteDataSource(apiService)
+	fun provideHomeRepository(homeSearchData:HomeSearchData): HomeRepository {
+		return HomeRepository(homeSearchData)
 	}
+
 	@Provides
 	@Singleton
-	fun provideRepository(remoteDataSource: RemoteDataSource): HomeRepository {
-		return HomeRepository(remoteDataSource)
+	fun provideDetailRepository(detailSearchData: DetailSearchData): DetailRepository {
+		return DetailRepository(detailSearchData)
 	}
 
 }
