+++ app/src/main/java/com/example/potatoservice/ui/detail/DetailSearchData.kt
@@ -1,0 +1,12 @@
+package com.example.potatoservice.ui.detail
+
+import com.example.potatoservice.model.remote.ActivityDetail
+
+interface DetailSearchData {
+	fun lookDetail(id: Int, callback: DetailCallback)
+
+	interface DetailCallback {
+		fun onLoaded(activityDetail: ActivityDetail)
+		fun onFailed()
+	}
+}
