+++ app/src/main/java/com/example/potatoservice/ui/detail/DetailViewModel.kt
@@ -7,7 +7,6 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.asLiveData
 import androidx.lifecycle.viewModelScope
-import com.example.potatoservice.ui.home.HomeRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
@@ -16,7 +15,7 @@ import javax.inject.Inject
 @HiltViewModel
 class DetailViewModel @Inject constructor(
 	application: Application,
-	private val repository: HomeRepository
+	private val repository: DetailRepository
 ): AndroidViewModel(application){
 	//리뷰 퍼센트지 값
 	var review1 = "50"
