+++ app/src/main/java/com/example/potatoservice/ui/home/HomeSearchData.kt
@@ -1,0 +1,14 @@
+package com.example.potatoservice.ui.home
+
+import com.example.potatoservice.model.remote.Activity
+import com.example.potatoservice.ui.share.Request
+
+interface HomeSearchData {
+	fun search(request: Request, callback: LoadCallback)
+
+	interface LoadCallback {
+		fun onLoaded(activities: List<Activity>)
+		fun onFailed()
+	}
+
+}
